+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Release notes follow the [keep a changelog](https://keepachangelog.com/en/1.1.0/
 ### Removed
 
 - AndroidManifest: The plugin no longer ships with a custom AndroidManifest.
+- Python SDK handling has been removed. This feature was only required for Unity 2019 on MacOS 12+
 
 ## [2.5.6] - 2025-02-11